requirements.txt
streamlit
streamlit-navigation-bar
openai
python-dotenv